requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.10
django-cors-headers==3.2.1
django-grappelli==2.14.1
django-import-export==2.1.0
django-jet==1.0.8
django-rest-auth==0.9.5
djangorestframework==3.11.0
et-xmlfile==1.0.1
google-api-python-client==1.4.1
httplib2==0.17.0
idna==2.9
jdcal==1.4.1
MarkupPy==1.14
oauth2client==4.1.3
odfpy==1.4.1
openpyxl==2.6.4
Pillow==7.0.0
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
rsa==4.0
six==1.14.0
sqlparse==0.3.1
tablib==1.1.0
uritemplate==3.0.1
urllib3==1.25.9
xlrd==1.2.0
xlwt==1.3.0
